Add service group to enable state report
This patch enables higgins services, such as
higgins-conductor, to report their state
periodically.

While the service is starting, a periodical
task is created and it reports the information
of the service to Database.

To complete the unit test for service group,
some necessary test code is added.

Implements: blueprint support-service-group
Change-Id: Id3dff15b46feb5874ffec3231bfa8d526ac44cf1