Set the scope of the check for unique container names
If a scope for the name check is set, launch of a new
container with a duplicate name will result in an
''ContainerAlreadyExists'' error. The uniqueness is
case-insensitive. Setting this option can increase the
usability for end users as they don't have to distinguish
among containers with the same name by their IDs. The concept
is currently used in nova, i borrowed it from there after
finding the difficulty in using commands with ids.
Possible values:
* '': An empty value means that no uniqueness check is done
  and duplicate names are possible.
* "project": The container name check is done only for
  containers within the same project.
* "global": The container name check is done for all containers
  regardless of the project.

Change-Id: I71e86f67dfe41340061ac6b956425b5dcd9295e0