Remove dict-compat from the image object
This patch removes the need of having dictcompat from
the image object. Luckily we have around 5 objects in
zun that would need this changes.

DictCompat is not really needed, since that's the old
way of referencing an object field just like a dictionary
object in python.

Currently a better way needs to be implemented to avoid
conversion of the image object to a dict
back and forth. This could be achieved if we implement
the API validation blueprint.

Change-Id: Ia0f5a4415b70934e7d47d77594f0d43ae6cb65bc
Partially-Implements: blueprint rm-object-dict-compat