Fix Container name match regular expression
Currently container name may contain spaces, the
regular expression is not valid. This patch fixes that
regular expression and adds unit tests.

NOTE: This patch doesn't fix the regular expression for
image name, because that is more complicated to be handled
now, i will propose another patch with that fix.

Close-bug: #1640680

Change-Id: Id99afda8dab5ee2f86d6f50e1988e9eba8542270