Implements image_pull_policy
This patch implements image_pull_policy, which determines if
the image should be pulled prior to starting the container.

It can have following values:
ifnotpresent: only pull the image if it does not already exist on the node.
always: Always pull the image from repositery.
never: never pull the image

If image_pull_policy is not specified then:
1. If the tag is latest, defaults image_pull_policy is always.
2. Otherwise default is ifnotpresent

Partial-Implements: blueprint image-pull-policy
Change-Id: If2715d7724fd5336fb25d380f125e1485ca302d0