zun run URL changed to /v1/containers?run=true
The "run" URL violating the OpenStack API guide.
This is the "run" API:

  POST /v1/containers/run

In particular, the guide said:
It is often the case that an API will have URIs that represent collections
for resources and individual members of that collection in a hierarchy.
Per my understanding, it means the following:
* /v1/containers represent collections of resources
* /v1/containers/<ID or NAME> represent individual members of that collection
However, the URL /v1/containers/run might be interpreted as an individual member
with name "run". This is not an effective URL design. According to the guide, it
suggested to change it as following

 POST /v1/containers?run=true

Closes-Bug: #1650820

Change-Id: I058a947afd4d4635fa8cc2f73ed2ff1fa25db13d