Rename compute/api.py to compute/rpcapi.py
The file contains code of RPC API so let's rename it to make it
clear. This also makes the naming pattern consistent with other
projects (i.e. Nova). Most importantly, I am going to create the
file compute/api.py for non-RPC compute API, so we have to free
the file name.

Change-Id: I1fbda7430b4f4ca3ca05ef1e1fde52ced590ae35