Added support for Auto remove the container
Docker run command has an option that automatically
removes the container when it exits. This commit introduces
the equivalent in Zun. In particular, this commit does
the followings:

* Introduce an 'auto_remove' field in container. If this
  field is set, the container will be automatically
  removed when it exits.
* Introduce a new container state 'Delete'. This states
  indicated that the container has been automatically
  removed, but the record is still existed in DB.
* Bump the docker REST API version to 1.25 or higher
  since 'auto_remove' feature was introduced in 1.25.

There are several future work of this commit. The first task is
to add a tempest test to verify the container is indeed
auto-removed in docker daemon. The second task is to introduce
a periodic task that purge container in 'Deleted' state.

Co-Authored-By: Hongbin Lu <hongbin.lu@huawei.com>
Co-Authored-By: Kien Nguyen <kiennt@vn.fujitsu.com>
Related-Bug: #1644901

Change-Id: Ic6d35274a49648bde5e0e7486453a6d1a13f6f2e