Remove Image and ZunService class from controllers
Earlier Image and Zunservice classes were used to validate
the request body for the Image and Services operations.
After moving the validation to JSON schema validator,
it has no major use case. So this patch removes the Image
and ZunService class and moves the logic to represent
images in to 'views' folder.

Partially-Implements: BP api-json-input-validation
Change-Id: Ie333bec64d5bd375886e1b2768b24d7db96b2dba