Introduce numa related objects
This patchset introduces the numa objects needed to represent the
CPUset related data in a standard NUMA object format.

This patch also introduces a driver method to query the host_cap
to figure out the cpuset information. Further improvments to this
patch may include detecting a linux operating system to decide
how the host_capability should be obtained.

This patch introduces a concept of floating_cpu_sets. The floating
cpusets are those, that are not meant for pinning. Such cpusets
are then excluded while doing the inventory of the host. These
cpusets are not the ones on which pinning is performed.

The floating cpuset is obtained by looking at a CONF parameter.
Further, the logic/format to specify the floating cpuset is kept
similar to the nova logic of obtaining the vcpu_pin_set and the
parsing logic has been more or less kept similar to nova/virt/
hardware.py get_vcpu_pin_set() method.

Change-Id: Iae0336a0996af1fa899d6ecae34d1bc3d881829e