Add "run" parameter to the exec API
The "run" parameter determines whether to run the command after
creating the exec instance. If "run" is true, Zun will create and
start the exec instance, and await for the execution of the command
to finish. Finally, inspect the instance to retrieve the status
of the command execution. If "run" is false, Zun will create an exec
instance without running the command. Instead, the ID of the exec
instance will return to users and the returned ID could be used
to trigger command execution at client side.

NOTE: We might add support for interactive mode in the future and
it requires Zun server to return the exec ID to zunclient, which
is a use case for setting "run" to false.

Change-Id: I05432fd150db89f80b14fab68012048c09d1cf36