Return docker url as well on interactive mode
This is a requirement for supporting interactive execute inside
a container. Generally speaking, this feature works by letting server
expose the URL that is used by client to connect and enter the
container. This patch is the first step by letting server return
the URL. There is another patch in zunclient for leveraging the URL
to connect to the container:

https://review.openstack.org/#/c/445234/

Note that it is insecure to directly expose such URL and this will
be fixed by another BP:

https://blueprints.launchpad.net/zun/+spec/stream-via-rest-api

Partial-Implements: blueprint support-interactive-mode
Change-Id: If987222ecf5419a305810858e304a4c2b4a78474