Allow host of docker daemon to be configurable
An immediate use case is to enable attaching to containers that
are behind the proxy. In this case, the ip address of docker
daemon won't be reachable from outside thus attaching is
impossible. With this config, cloud admins can set a reachable
ip address for external access of docker daemon.

Change-Id: If4981154139c98ad4a8373744d29938214991ca6