Avoiding setting container to error on list
It is possible that a container was created in DB but haven't
been created in docker yet. For example, a container is in the
process of creating or deleting. In before, we set container to
error state if it is not found in docker. This caused an incorrect
report of container state. This commit fixed it.

Change-Id: I51af16e7837de227aff3ee4783acf7fe5cce12ca
Closes-Bug: #1686490