Set 'addresses' field at an earlier stage
In before, we set the 'addresses' field of containers after we
mark the container as 'Created'. However, clients (i.e. Heat) might
try to retrieve the addresses right after the container is created.
As a result, an empty 'addresses' is returned which is suboptimal.

This commit move the setting of addresses before the first 'save'
so that users will get the addresses as long as the container is
of an active state.

Change-Id: I71d2f3d1108772fdffe7edf13f2d250daacc9eb4
Closes-Bug: #1686523