Modify the Cinder integration spec
Fuxi is a docker volume plugin that allows users to use Docker API
to manage data volumes backed by Cinder or Manila. Fuxi is a
desired solution for some use cases, but it has several
disadvantages:
* It currently doesn't support multi-tenancy. Fuxi-server will
  use a pre-created credential to authenticate againest
  Cinder/Manila. Therefore, it cannot manage resources that is
  outside of the tenant of the credential.
* It is currently under heavy development. In particular, Fuxi
  team is planning to port the code from Python to Golang [1].
  The stability of the project will be affected.
* It is a web service. That means extra overhead for operators
  to install and manage the extra service.

Due to the limitations, we proposes to introduce another driver
as an alternative. The new driver will be a standalone module
(instead of a service) that directly interfaces with Cinder.
The architecture will be designed to be pluggable so that
the support of multiple volume drivers is possible.

[1] https://blueprints.launchpad.net/fuxi/+spec/convert-to-golang

Co-Author-By: Kien Nguyen <kiennt@vn.fujitsu.com>
Partial-Implements: blueprint direct-cinder-integration
Change-Id: Idbe4511378dd321d3087f0abcde5fec01b31e774