Avoid creating port without security groups
If the security groups key is set and value is None or an empty list,
the created neutron port won't have any connectivity. This is a
pitfall. We should avoid setting the security group if it is not
provided by users. As a result, neutron will associate the 'default'
security group to the port.

Change-Id: I7d79bf2ddb8272f003a2f648532226f405822649
Closes-Bug: #1694336