Unset container's host if it failed to create
This is for fixing the wrong calculation of available resources
in compute host. Resource comsuption of a host is calculated by
listing all containers that has a 'host' field equal to current
host. If a container failed to create, it should has its 'host'
field unset so that its resource won't be attributed to the host.

Change-Id: Id095e7b02f6a6f45c15f97f029027dfbc3955b7e