Optimize disconnection of container from network
The disconnection of container from network needs to delete all
the neutron ports of the container. In before, we searched
neutron ports by device_id/EndpointID. The problem is EndpointID
is not available if the container is not running. This commit
retrieves neutron ports from the 'addresses' field of the container
instead.

Partial-Implements: blueprint make-sandbox-optional
Change-Id: Ibe9383d735640d66cc9f24ce3caccd281a64b4e7