Optimize add_security_groups_to_container
The logic to add security groups to container is to search all
the neutron ports of the container and update the security_groups
fields of those ports. In before, we searched neutron ports
by device_id/EndpointID. The problem is EndpointID is not available
if the container is not running. This commit retrieves neutron ports
from the 'addresses' field of the container instead.

Partial-Implements: blueprint make-sandbox-optional
Change-Id: I845872665a766cae11528f226350399497d58b64