Set sandbox id immediately after sandbox create
When sandbox created and start to set the sandbox_id, there is a
chance that start the sandbox failed, then the sandbox id is not
set to the container metadata. When delete container, even by -f
still cannot delete the sandbox as there is no sandbox id in the
container metadata.

Set the sandbox id to container immediately after sandbox created.

Change-Id: Iebbf2480d5c1332644e4081bc54003101a7ac1ed
Closes-Bug: #1699635