Optimize helpinfo in zun/conf
Remove function _() from zun/conf. Currently, some configs helpinfo
has translation and the others do not, which is inconsistent.
There are two options to resolve this. The first option is to remove
all translation from config. The second option is to enforce
translation for all configs. This commit suggested the first option
because configs are not directly exposed to end-users so the value
of translated configs helpinfo is not too much.

Change-Id: Iadd8333b5156696477b4dced01077e77fa42afb2