Implement create with existing neutron port
Support creating container with existing neutron port. Users
can pass the uuid/name of an existing neutron port via the 'nets'
parameter. If port is specified, Zun will use the existing port
instead of creating a new port.

If a container is deleted, its neutron port will be deleted regardless
of existing or not. A future work is to determine if the port is
pre-existing and avoid deleting existing port.

Closes-Bug: #1702581
Change-Id: I204b4bbe3de9e275690c4a64acdfe557f134b7ee