Fix hard code kuryr driver name bug
kuryr driver name is hard coded in zun, if user installed it by
docker plugin install kuryr/libnetwork2, the driver name maybe the
'kuryr/libnetwork2:latest', which will cause create docker fail.

Make the dirver name passed to docker as an conf options, which
default is 'kuryr' as the default driver is kuryr.

Change-Id: I9cc52585950bc44889a3e468476599e27007e8d0
Closes-Bug: #1703383