Move get_available_network to API layer
In before, if there is no network available, users don't know what is
wrong until the container is scheduled to a compute node and fail.
This commit moves the searching of neutron net to zun-api so that
users will receive an error response earlier if the request is
invalid.

Change-Id: Ic9fd47c6e4aff4131ebaad73a3bed0f51eeb118c
Related-Bug: #1702581