Fix the 'addresses' field of container
The 'addresses' field is a map of ip addresses. In before, it
used docker network name as key, which is hard to be interpreted
by users, since docker network is hiden from end-users. This commit
switches to use neutron network uuid as key. A use case is from our
Heat plugin, which requires neutron network information to do
orchestration.

Change-Id: Ic3e4ee80962cdd777543e5c9c988fc0cdc623dad