Consolidate container_create and container_run
The code of container_create and container_run are duplicated in
controller, zun api, rpc api, and compute manager. That is because
container_run is just a special case of container_create.
Consolidate these codes will be beneficial for future. In particular,
the cinder integration implementation will take advantage of this.

Change-Id: I6d090e99c3c1be80a2aa7a96137cf6afc004d466