Introduce DB model for volume mapping
We are going to introduce a concept 'volume mapping' that represents
a mapping between Cinder volume and Zun container. Each volume mapping
entry stores the uuid of the Cinder volume and Zun container, as well
as other necessary information for volume bind mounting.

In the follow-up patches, we are going to allow users to create a
container with cinder volume(s) bind-mounted. Zun will call
Cinder API to attach the volume to the compute node, then mount
it to the container. The volume mapping table will be looked
up for various purposes, such as look up the volumes by a container
uuid, store and look up the connection_info, etc.

Change-Id: Id11246a16667bac407a0e4255437b84755a75784
Partial-Implements: blueprint direct-cinder-integration