Introduce utility for mounting Linux device
This utility is needed for bind mounting volume. Assume
a Cinder volume is created and attached to the compute node as
a Linux device. The next step is to mount the Linux device to
a path in local file system so that Docker can bind mount it
to containers.

The utility module is basically a wrapper to execute a few
Linux commands to mount, unmount, and read mounts in local host.
The mounting device should be a block device backed by Cinder.
If the mounting device is unformatted (the Cinder volume is newly
created), the module will try to format the device. Otherwise,
the device won't be formatted.

Change-Id: Icd9c44223989d8b5582d7b95dcab21dab101c8e8
Partial-Implements: blueprint direct-cinder-integration