Implement APIs for mounting Cinder volumes
Add a new attribute 'mounts' to the request to create a container.
Users can use this attribute to specify one or multiple mounts for
the container. Each mount could specify the source and destination.
The source is the Cinder volume id or name, and the destination is
the path where the file or directory will be mounted in the container.
For examples:

  [{'source': 'my-vol', 'destination': '/data'}]

Partial-Implements: blueprint direct-cinder-integration
Change-Id: Ic11c0bba499ce107a7fab35059671c3d4254d343