Remove network_standalone variable
This is a minor refactoring of the driver code. The variable
network_standalone is used for testing if the driver should
provision and setup the network. However, it is possible to
achieve the same without using this variable. This patch proposes
to remove this variable to make the code look better.

Partial-Implements: blueprint direct-cinder-integration
Change-Id: Ic8f298c10bd4ed74e04c8ec5c2974abb64b8391d