Bring hostname option back
Zun already had this option before, but had to remove due to the
Conflict Exception (relate to sandbox). At this time, we made sandbox
optional. Therefore, if use_sandbox option is False, we can use
hostname option. If it is True, raise the Exception to let user know.

This option will help us to implement the
blueprint internal-dns-resolution.

Change-Id: I2f8826fc66e94c0d4452ff32e3bc17700ef6b668
Closes-Bug: #1716849