Use devstack gate to run tempest tests
In before, we run tempest tests manually at post_test_hook.sh, which
is suboptimal and hard to maintain. This incurs challenges on splitting
tempest tests as well.

Most projects use devstack gate to run tempest tests. It is done by
setting DEVSTACK_GATE_TEMPEST_REGEX to point to the right set of
tests. Let's switch to this approach.

Change-Id: If29151755fb958d308ec3759b5d1a2a1200ccfc5