Delete containers created in tests only
In before, tempest tests will delete all containers on teardown.
This is problematic if concurrency is set larger than 1 since there
would be some race conditions when multiple processes are trying
to do the cleanup.

This patch proposes to record the list of containers created in
tests and cleanup those containers only.

Change-Id: I0823fbbf999f346f08d283b76073a6da461bb9e3