Implement basic policy module in code
This change prepares the zun project to start implementing policies
in code. Subsequent patches will register more zun policies in code
and remove the corresponding entry from the policy file maintained in
source.

This is part of a community effort to provide better user experience
for those having to maintain RBAC policy. More information on this
effort can be found below:

  https://governance.openstack.org/tc/goals/queens/policy-in-code.html

bp policy-and-docs-in-code

Change-Id: I4b0cf1203f7d7ddda240106bd39ef35bce604810