Move inventory host to compute manager
This is because this task updates the resource tracker instance
and this resource instance is duplicated with the one in compute
manager. The duplication of resource trackers is problematic on
tracking resources. For example, PCI tracking won't work as
expected.

This patch move the task to compute manager so that it can
share the same resource tracker with compute manager. The
duplication of tracking is removed.

Change-Id: Ieb125a239b5e9f3287701f1d30f35171886e14e1