Container action response 500
Zun action APIs (i.e. exec, top, etc.) will return 500 if Docker
return 409 (Conflict). Below are observations:

* Docker will validate container state before executing a specific
action, and raise state conflict (409) error if the state of the
container doesn't allow the action to perform. For example,
'exec' is not allowed if container is on state 'Stopped'.

* Zun will do a similar state validation in zun-api.
* Zun will sync the container states from Docker every 60 seconds.

As a result, the state in Zun and Docker might be different during
the 60 seconds interval. This 500 error happened if the state between
Zun and Docker is out-of-sync (sometimes in between the 60 seconds).
This makes the state validation passed in Zun but failed in Docker.

Change-Id: If1db262881ac89c62d5a2e7f3cfe67a09724ef5a
Closes-Bug: #1723499