Change the Flake8 ignore list
This makes our ignore list consistent with the OpenStack cookiecutter,
which is the suggested config to get started. We might further refine
the list later. In addition, due to the change of ignore list, several
disabled rules are enabled now, so we need to fix the style of the code
in the same patch.

Change-Id: I18ebc915678eee40f791e7f10ab428f6796e9550