Set network mode on creating container
This patch changed the workflow to configure the network in
docker. The old workflow is as following:
* Create container with default network_mode (which connected
  to 'bridge' network by default)
* Disconnect from 'bridge' network
* Connect to each requested neutron network

The problem is that docker-py doesn't support connect with
specified MAC address [1]. This caused problems on SR-IOV support,
which required to set the MAC address. Therefore, we need to change
to a new workflow:
* Create container with the first requested neutron network
* Connect to the rest of the neutron networks

[1] https://github.com/docker/docker-py/pull/1756

Change-Id: Ib26fc3458c95c3b3d5df9566acdb0a2d28d1eb30