Allow duplicated container name by default
In before, the default config disallow duplicated name within
tenant. This is a bad default because zun auto-generate container
name if it is not specified. It is possible that the same name
is generated twice which causes failure. This makes our API
unpredictable which is not a good practice. Therefore, we changed
the default to allow duplicated name.

Change-Id: Ic0f4d1279c0822aceac9e1983dc500afe9834a59