Create cinder volume in the correct tenant
If we use a user with admin privilege to create container with
a dynamically created cinder volume, the cinder volume won't be
created in user's tenant (it will actually be created in 'service'
tenant). This is due to a bug on keystone auth creation in
commom/keystone.py module. This patch fixed it.

In addition, this patch also adds the check on the dynamically
created cinder volume to ensure it is usable.

Change-Id: I312e643d02dc5a188d03c50cbac73c2342a61a82
Closes-Bug: #1740938