Avoid calling _fail_container in some cases
This is because _fail_container will set the container to Error
state in DB and do the necessary cleanup. However, the docker
container might be still running in Docker. We treat the states
in Docker as the source of truth and periodically sync Docker
container's states to Zun DB. Therefore, bypassing Docker and
directly changing container's state in DB is bad and might
lead to state inconsistency between Zun and Docker.

The side effect of this patch is that users cannot query the
result of each action from the container's resource. Users are
recommended to use the new action APIs [1] to achieve the
equivalent.

[1] https://review.openstack.org/#/c/548853/

Depends-On: Ic3fa5e879baa6a5a01a1ae83740a2149a6f8003b
Change-Id: I77346ef302ac7c641c798b734e18eac55dacc8bc