Reraise exception on creating container
It is desirable to raise exceptions to upper stack so that it
can be captured and handled. For example, we are introducing
action APIs that will record exception details if there is
an exception that has been raised to the upper stack.
Otherwise, the exception information will lose and the action
will be wrongly recorded as success.

Change-Id: I2737102510e04383c878c47441f46a86fed9d744