Temporarily relax validation of 'mounts'
We currently restrict the 'mounts' parameter to be one of the
following forms:
1. {'source': xxx, 'destination': xxx}
2. {'size': xxx, 'destination': xxx}

However, the python-zunclient sends the following:
i. {'source': xxx, 'destination': xxx, 'size': ''}
ii. {'size': xxx, 'destination': xxx, 'source': ''}

Either i and ii is considered to be invalid because it matches
both #1 and #2. However, the valid request should matches either
The long term fix is to make the python-zunclient stop sending
empty string to server.

Change-Id: Id8ffc5d783923c2783565b6d019b250feae2031d
Partial-Bug: #1747282