api-ref: remove query parameters from URL
By looking at the api-ref of other openstack projects, they don't
list the query parameters in the URL. Actually, the type of each
request's parameter has already been documented in the table so
it is necessary to list them in URL as well. Doing that will
make the API doc looks verbose.

Change-Id: I0e8601a5e8ec8d537bfafec6e59e16cf1c48aac3