Add availability_zone to service
Add a new attribute 'availability_zone' to the 'service' resource.
This attribute indicates the availability zone of the zun-compute
agent. Cloud administrators can customize the availability zone of
each compute node by setting the following in the config file:

  [DEFAULT]
  default_availability_zone = my-zone

If this is not set, the default availability zone is 'nova'
(we use this value for consistency with neutron and cinder's defaults).

In the future, we will add support for scheduling containers to
a specified AZ and the 'availability_zone' attribute will be used
by scheduler to check if the availability zone of the compute node
equals to the one requested by API users.

Change-Id: I4a4206b4eb0aa5149bbfc8ab72ae408a08317de4
Partial-Implements: blueprint zun-availability-zone