Support running containers in specified AZ
Add a new parameter 'availaibility_zone' to the REST API for
creating containers. The API server will pass down this parameter
the the scheduler as extra_spec. In the filter scheduler, add
a 'AvailabilityZoneFilter' filter for selecting hosts in the
requested availability zone. Below is the host selection scenario:

* If users specify an AZ for the container, scheduler will choose
  a host in the user-specified AZ.
* If users don't specify an AZ for the container, the AZ will be
  default to CONF.default_schedule_zone if this config is set.
* If users don't specify an AZ for the container and
  CONF.default_schedule_zone is not set (None), the scheduler will
  choose a host in arbitrary AZ.

Partial-Implements: blueprint zun-availability-zone
Change-Id: I01b8069f76e87a8e0c782da0888eee8a8cd58193