Api-ref: add a warning to execute_resize endpoint
In the API reference, we'd better to provide more information
about the execute_resize API endpoint:
https://developer.openstack.org/api-ref/application-container/#resize-tty-when-execute-command-in-a-container
This endpoint is not designed for end-users. It is designed
for UI or CLI to adjust the TTY's size of a command
execution in a container. Consider documenting this
with a warning message (as what we did for the resize
endpoint: https://developer.openstack.org/api-ref/application-container/#resize-a-container).

Change-Id: I87f33d0d20a1b73f46a67715930b7ac86e256756
Closes-Bug: #1749234