Get local containers list from driver
In before, we get the list of local containers from database.
In particular, we perform a DB call to list containers with
filters that matches the container's host to current host.
There are some corner cases that could produce an incomplete
list of local containers. For example, if a container is present
in the driver (Docker) but couldn't be found on querying the DB
due to host mismatch (which could happen for reasons such as
the host of the compute node is misconfigured).

To resolve the issue, this patch proposes to get a list of
containers from DB and get another list of containers from driver.
Then, take a union set of both lists. The final list will present
to periodic task to sync the container's state.

Change-Id: I8279634e78a69419d5b18751c9f125335bd66d5a