Add unit test for filter with list
In before, we supports filter containers with string value.
For example, the filter can be

  {'name': 'my_container'}

The result will return a list of containers with that name.
Right now, we add support for list as filter value.
For example, the filter can be

  {'name': ['my_container', 'my_second_container']}

The result will be a list of containers with the name in the list.
This patch adds unit test case for this feature.

Change-Id: Ia583934ef0d3e39b043eeb059a5aa6ce58ec3503