Pause container before doing commit
Docker daemon will pause the container being committed if we don't
pause the container. However, this will lead to inconsistency
between Docker daemon and DB (the container's state is Running in
DB but Paused in Docker). Although the inconsistency will be
converged in periodic task, we can do better by explicitly pausing
the container before doing the commit and unpause the container
after. This will keep the container's state consistent during and
after commit and avoid bug like #1749879.

Change-Id: Ie872e71e386714b791f3277fcdb2326e09235128
Related-Bug: #1749879