Expose container filters via API
Right now, in DB layer, we support to filter containers
with certain attributes. In particular, we support
filter containers by attributes such as name, image,
project_id, status, etc. [1]. It is better to expose this
filter via REST API level. For example:

  GET /v1/containers?project_id=<my_project_id>

This should return a list of containers in particular project.

[1] https://github.com/openstack/zun/blob/master/zun/db/sqlalchemy/api.py#L132

Change-Id: I599a9f48a966d564da718b97bdad7cfac6ae4dfe
Closes-Bug: #1750208