make glance driver support image:tag
Currently, glance does not support processing tag.
The purpose of the BP is to enable glance_driver to support
the processing of tag

Support the user to enter a glance image like: image:tag

1/ When the user uploads a docker image to glance,
the image must be named in the way of 'image:tag',
that means, the 'image:tag' is a image name.

2/ When the user uploads docker image to glance,
the image needs to set the tags parameter.
eg. 'glance image-create --tags tag:test --file ubuntu.tar
--name ubuntu --disk-format qcow2 --container-format docker --progress'

Currently implementing the second scheme

Change-Id: Ie44f4605512552d2c204c4b410f1a3bce0f2f3c0
Implements:blueprint glance-support-tag