Fix an issue of command population
Docker is returning a list representation of command and Zun is
currently using a string representation of command. In before,
we convert list to string like below:

  ["nginx", "-g", "daemon off;"] -> 'nginx -g daemon off;'

The problem is that the string representation is ambiguity as
we lose the information about how the original command were
tokenize. This causes problems of container rebuild, which is
a new feature to rebuild the container with original information.

This patch fixes this issue by quoting each token. The conversion
will look like:

  ["nginx", "-g", "daemon off;"] -> '"nginx" "-g" "daemon off;"'

Change-Id: I9eede500d0fbefe4b62d55ee3c9b6ab2facc36f0