Move pull_image to container driver
This refactoring is for making the code structure more consistent.
In particular, the image driver will be decoupled from compute
manager and align with volume/network driver.
In the new structure, compute manager is mainly for interfacing
with docker driver and docker driver is for interfacing with
image/volume/network drivers.

Change-Id: Ie5bfbe6dac36818348985b9bdc11106a04f77b8c