Move search_image to container driver
This refactoring is for making the code structure more consistent.
In particular, the image driver will be decoupled from compute
manager and align with volume/network driver.
In the new structure, compute manager is mainly for interfacing
with docker driver and docker driver is for interfacing with
image/volume/network drivers.

Change-Id: Ia8e6c9d8868fde1f650f9c56cade5e4b7dde507b