Move create_image and upload_image_data
This refactoring is for making the code structure more consistent.
In particular, the image driver will be decoupled from compute
manager and align with volume/network driver.
In the new structure, compute manager is mainly for interfacing
with docker driver and docker driver is for interfacing with
image/volume/network drivers.

Change-Id: Ib270062e3a274d2c5d98c05375520ac46b644984