Synchronize on network attach/detach
There are race conditions on concurrent network attach/detach
requests on the same container. We need to lock down the container
before doing these operations. Due to this change, the RPC call
will be blocking so we switch from RPC call to RPC cast so that
it won't block the API thread.

Change-Id: Ieca5bdfe5ca427f30c889bf5fc82c49a104d1477