Add a note about configuration of ws url
The common problems that users ran into is that they are not
able to access the container's console in browser (via Horizon).
The reason is that the ws_proxy's URL is not accessible from
the browser. Either the configured IP address is not accessible
from outside or the port was blocked. This patch adds a note to
explain how to configure the websocket proxy's URL propertly.

Change-Id: Ifac45550f22409c2348e1b14ff46606a7bf09a89
(cherry picked from commit e5a9190e1c669e035bb350fb1cb1bf474c668cb4)