Use policy for control the 'host' field
In before, we manually remove the 'host' field from container
if the request was not made by admin users. This patch allows
operators to define a policy to decide whether to expose
this field. This is more flexible.

In addition, the field removal logic is moved to 'format_container'
to ensure that the field was removed (by policy) in all HTTP
methods (instead of 'GET' method only).

The policy is of the form "container:get_one:<field_name>".
For example, if the field is 'host', the corresponding policy
would be "container:get_one:host". If there is no policy defined
for a field, this field is allowed. Otherwise, the visibility
of this field is decided by the defined policy. The policy
enforcement logic is extended to allow a field 'might_not_exist'.

Change-Id: I40a0bfb4c26b8b46218f36ae0da9043b0f40b10b